+++ config/notifiers.go
@@ -18,6 +18,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/pkg/errors"
 	commoncfg "github.com/prometheus/common/config"
 )
 
@@ -468,17 +469,38 @@ type OpsGenieConfig struct {
 	Description string            `yaml:"description,omitempty" json:"description,omitempty"`
 	Source      string            `yaml:"source,omitempty" json:"source,omitempty"`
 	Details     map[string]string `yaml:"details,omitempty" json:"details,omitempty"`
-	Teams       string            `yaml:"teams,omitempty" json:"teams,omitempty"`
-	Tags        string            `yaml:"tags,omitempty" json:"tags,omitempty"`
-	Note        string            `yaml:"note,omitempty" json:"note,omitempty"`
-	Priority    string            `yaml:"priority,omitempty" json:"priority,omitempty"`
+	// Teams field is deprecated.
+	Teams      string                    `yaml:"teams,omitempty" json:"teams,omitempty"`
+	Responders []OpsGenieConfigResponder `yaml:"responders,omitempty" json:"responders,omitempty"`
+	Tags       string                    `yaml:"tags,omitempty" json:"tags,omitempty"`
+	Note       string                    `yaml:"note,omitempty" json:"note,omitempty"`
+	Priority   string                    `yaml:"priority,omitempty" json:"priority,omitempty"`
 }
 
 // UnmarshalYAML implements the yaml.Unmarshaler interface.
 func (c *OpsGenieConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	*c = DefaultOpsGenieConfig
 	type plain OpsGenieConfig
-	return unmarshal((*plain)(c))
+	if err := unmarshal((*plain)(c)); err != nil {
+		return err
+	}
+
+	if c.Teams != "" {
+		// Should we allow this for backward compatibility? Should we fill responders from this?
+		return errors.New("teams field is deprecated in Opsgenie config. Use responders instead.")
+	}
+
+	return nil
+}
+
+type OpsGenieConfigResponder struct {
+	// One of those 3 should be filled.
+	ID       string `yaml:"id,omitempty" json:"id,omitempty"`
+	Name     string `yaml:"name,omitempty" json:"name,omitempty"`
+	Username string `yaml:"username,omitempty" json:"username,omitempty"`
+
+	// team, user, escalation, schedule etc.
+	Type string `yaml:"type,omitempty" json:"type,omitempty"`
 }
 
 // VictorOpsConfig configures notifications via VictorOps.